refactor: extract papr-core crate (UI-free data + ingestion)
Split the Tauri-free data layer, ingestion building blocks, sanitization
and OPML into a new workspace crate so both the desktop app and a future
agent CLI link the exact same code.

- New workspace at repo root; members: src-tauri, crates/papr-core, crates/papr-cli
- Moved error/models/db/sanitize/extraction/opml + ingestion{discovery,fetch,
  newsletter,parse,sources} into papr-core
- Tauri-coupled refresh scheduler relocated to papr_lib::scheduler
- src-tauri re-exports papr_core modules under their original crate paths,
  so commands/sync/translate/tray are unchanged
- papr-cli: stub binary (implemented next)

184 papr-core tests pass; app builds.
