+++ Cargo.toml
@@ -1,0 +1,37 @@
+[workspace]
+members = ["src-tauri", "crates/papr-core", "crates/papr-cli"]
+resolver = "2"
+
+# Shared dependency versions, inherited by the member crates via
+# `<dep>.workspace = true`. Keeping the data/ingestion stack pinned here means
+# the desktop app and the agent CLI link the exact same code.
+[workspace.dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+rusqlite = { version = "0.32", features = ["bundled", "functions"] }
+rusqlite_migration = "1"
+feed-rs = "2"
+reqwest = { version = "0.12", features = ["gzip", "brotli", "json"] }
+tokio = { version = "1", features = ["full"] }
+scraper = "0.21"
+ego-tree = "0.9"
+url = "2"
+opml = "1"
+encoding_rs = "0.8"
+imap = { version = "=3.0.0-alpha.15", default-features = false, features = ["rustls-tls"] }
+mail-parser = "0.11"
+ammonia = "4.1.2"
+dom_smoothie = "0.17"
+lol_html = "2"
+chrono = { version = "0.4", features = ["serde"] }
+anyhow = "1"
+thiserror = "2"
+log = "0.4"
+
+# The desktop bundle is size-optimized and stripped; the agent CLI inherits the
+# same release profile (one static binary to drop on an agent's PATH).
+[profile.release]
+codegen-units = 1
+lto = true
+opt-level = "s"
+strip = true
