+++ crates/papr-cli/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "papr-cli"
+version = "0.10.0"
+edition = "2021"
+description = "papr — an Agent eXperience Interface (AXI) over your Papr RSS feeds: read, search, triage and refresh from the shell."
+
+[[bin]]
+name = "papr"
+path = "src/main.rs"
+
+[dependencies]
+papr-core = { path = "../papr-core" }
+serde.workspace = true
+serde_json.workspace = true
+rusqlite.workspace = true
+reqwest.workspace = true
+tokio.workspace = true
+chrono.workspace = true
