+++ crates/papr-cli/src/main.rs
@@ -1,0 +1,4 @@
+fn main() {
+    // Placeholder — the AXI command surface is implemented in a later step.
+    eprintln!("papr-cli: not yet implemented");
+}
