+++ crates/papr-core/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "papr-core"
+version = "0.10.0"
+edition = "2021"
+description = "Papr's UI-free core: SQLite data layer, feed ingestion, and content sanitization. Shared by the desktop app and the agent CLI."
+
+[lib]
+name = "papr_core"
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
+rusqlite.workspace = true
+rusqlite_migration.workspace = true
+feed-rs.workspace = true
+reqwest.workspace = true
+scraper.workspace = true
+ego-tree.workspace = true
+url.workspace = true
+opml.workspace = true
+encoding_rs.workspace = true
+imap.workspace = true
+mail-parser.workspace = true
+ammonia.workspace = true
+dom_smoothie.workspace = true
+lol_html.workspace = true
+chrono.workspace = true
+anyhow.workspace = true
+thiserror.workspace = true
