+++ crates/papr-core/src/error.rs
