+++ crates/papr-core/src/ingestion/fetch.rs
