+++ crates/papr-core/src/ingestion/newsletter.rs
