+++ src-tauri/Cargo.toml
@@ -33,6 +33,10 @@ tauri-plugin-process = "2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 
+# The UI-free core: SQLite data layer, ingestion building blocks, sanitization.
+# Shared verbatim with the agent CLI so both link the same data code.
+papr-core = { path = "../crates/papr-core" }
+
 # data layer
 rusqlite = { version = "0.32", features = ["bundled", "functions"] }
 rusqlite_migration = "1"
@@ -81,9 +85,3 @@ log = "0.4"
 [target.'cfg(target_os = "macos")'.dependencies]
 objc2 = "0.6"
 objc2-foundation = "0.3"
-
-[profile.release]
-codegen-units = 1
-lto = true
-opt-level = "s"
-strip = true
