+++ src-tauri/src/commands.rs
@@ -8,8 +8,9 @@ use crate::extraction;
 use crate::ingestion::discovery::{self, DiscoveryResult};
 use crate::ingestion::newsletter::{self, NewsletterConfig};
 use crate::ingestion::sources::{self, Normalized};
-use crate::ingestion::{fetch, parse, scheduler};
+use crate::ingestion::{fetch, parse};
 use crate::models::*;
+use crate::scheduler;
 use crate::opml;
 use crate::sanitize;
 use crate::state::AppState;
