--- src-tauri/src/ingestion/mod.rs
@@ -1,8 +1,0 @@
-//! Ingestion: fetching feeds over HTTP, parsing them, and the refresh scheduler.
-
-pub mod discovery;
-pub mod fetch;
-pub mod newsletter;
-pub mod parse;
-pub mod scheduler;
-pub mod sources;
